+++ src/components/FormCadastroEmbarcacao/index.css
@@ -4,63 +4,71 @@ html {
   height: 100%;
   width: 100%;
   font-family: "Montserrat", sans-serif;
+  min-width: 100%;
+  min-height: 100%;
 }
 
 
 .container-cadastro-embarcacoes {
     width: 100%;
-    /* max-width: 900px; */
-    background: green;
+    /* background: green; */
     backdrop-filter: blur(10px);
-    border: 2px solid rgba(7, 7, 7, 0.2);
+    /* border: 2px solid rgba(7, 7, 7, 0.2); */
     border-radius: 10px;
-    padding: 40px;
+    /* padding: 40px; */
     display: flex;
     flex-wrap: wrap; /* Permite que as seções se organizem em linhas */
     justify-content: space-between;
     gap: 20px;
-    height: 95vh;
+    height: 55vh;
   }
   
   
 .container-cadastro-embarcacoes h1 {
       text-align: center;
       font-size: 30px;
-      color: #f1eded;
+      color: black;
     }
   .form-section-embarcacoes{
     width: 100%;
     display: flex;
+    margin: 0 15vw;
+    height: 100%;
   }
 
   .input-field  {
-    width: 60%;
-    height: 30%;
-    background: transparent;
+    width: 100%;
+    height: 100%;
+    /* background: red; */
     outline: none;
-    border: 2px solid rgba(255, 255, 255, 0.2);
-    border-radius: 40px;
+    /* border: 2px solid rgba(255, 255, 255, 0.2); */
     font-size: 16px;
     color: #fff;
-    padding: 20px 45px 20px 25px;
+    display: flex;
+    flex-flow: row wrap;
+    /* row-gap: 5vh; */
+    column-gap: 1vw;
+    justify-content: space-evenly;
+    align-items: center;
+
   }
 
  
   
   .input-field input {
-    width: 100%;
-    height: 100%;
-    background: transparent;
+    width: 25vw;
+    height: 7.5vh;
+    /* background: rebeccapurple; */
     outline: none;
-    border: 2px solid rgba(255, 255, 255, 0.2);
+    border: 1px solid rgba(19,141,147,255);
     border-radius: 40px;
     font-size: 16px;
-    color: #fff;
+    color: black;
     padding: 20px 45px 20px 25px;
   }
   
   .input-field input::placeholder {
-    color: #f0e9e9;
+    color: black;
   }
