+++ src/components/FormCadastroPerfilEmbarcacao/index.css
@@ -31,7 +31,7 @@
 .titulo-primeiro-embarcacao
 {
   color: black;
-  /* padding: 15px; */
+ 
 
 }
 .titulo-embarcacao {
