+++ src/pages/cadastroLocador/index.css
@@ -1,0 +1,60 @@
+html {
+  margin: 0;
+  padding: 0;
+  height: 100%;
+  width: 100%;
+  font-family: "Montserrat", sans-serif;
+  min-width: 100%;
+  min-height: 100%;
+}
+
+.container-cadastro{
+  width: 100%;
+  height: 90vh;
+}
+
+.titulo-cadastro {
+  padding-left: 7%;
+  padding-right: 7%;
+}
+.titulo-cadastro h2{
+  color: black;
+}
+
+.linhaHr{
+  border-top: 4px solid rgba(19,141,147,255);
+
+}
+
+.icones-embarcacao {
+  /* margin-top: 25px; */
+  height: 25vh;
+  display: flex;
+  column-gap: 5vw;
+  /* margin-left: 380px; */
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  padding: 3vh 10vw;
+  
+}
+
+.btn-icone{
+  width: 14%;
+  height: 95%;
+  border: 1px solid rgba(19,141,147,255);
+  border-radius: 13px;
+  background-color: rgba(245, 245, 245, 0.79);
+}
+.btn-icone :hover{
+/* transform: scale(2); */
+scale: 2;
+}
+.icones-svg img{
+  width: 70px;
+  height: 70px;
+  
+
+}
+
+
