+++ src/router/routes.jsx
@@ -2,8 +2,9 @@ import { createBrowserRouter } from "react-router-dom";
 import Login from "/src/pages/login/Login.jsx";
 import Home from "/src/pages/home/Home.jsx";
 import RecuperacaoSenha from "../components/FormRecuperacaoSenha";
-import CadastroEmbarcacoes from "../pages/cadastro/CadastroEmbarcacoes";
 import CadastroDescricao from "../components/FormCadastroDescricao";
+import CadastroEmbarcacoes from "../components/FormCadastroEmbarcacao";
+import CadastroLocador from "../pages/cadastroLocador";
 
 
 
@@ -13,8 +14,9 @@ import CadastroDescricao from "../components/FormCadastroDescricao";
 
 const router = createBrowserRouter([
     {path: "/", element: <Login />},
-    {path: "/cadastroEmbarcacoes", element: <CadastroEmbarcacoes/>},
+    {path: "/cadastroLocador", element: <CadastroLocador/>},
     {path: "/cadastroD", element: <CadastroDescricao/>},
+    {path: "/cadastroEmbarcacoes", element: <CadastroEmbarcacoes/>},
     { path: "/recuperacaoSenha", element: <RecuperacaoSenha /> },
     { path: "/home", element: <Home /> },
